separate sessions into own POM
