+++ test/e2e/pages/variables.ts
@@ -7,7 +7,7 @@
 import { Code } from '../infra/code';
 import * as os from 'os';
 import test, { expect, Locator } from '@playwright/test';
-import { SessionName } from '../infra';
+import { SessionName } from './session';
 
 interface FlatVariables {
 	value: string;
