+++ test/e2e/tests/console/console-sessions.test.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect } from '@playwright/test';
-import { SessionName } from '../../infra';
 import { test, tags } from '../_test.setup';
+import { SessionName } from '../../pages/session';
 
 const pythonSession: SessionName = {
 	language: 'Python',
@@ -32,102 +32,104 @@ test.describe('Console: Sessions', {
 		await app.workbench.variables.togglePane('hide');
 	});
 
-	test('Validate state between sessions (active, idle, disconnect) ', async function ({ app, interpreter }) {
-		const console = app.workbench.console;
+	test('Validate state between sessions (active, idle, disconnect) ', async function ({ app }) {
+		const session = app.workbench.session;
 
 		// Start Python session
 		await app.workbench.console.startSession({ ...pythonSession, waitForReady: false });
 
 		// Verify Python session is visible and transitions from active --> idle
-		await console.session.checkStatus(pythonSession, 'active');
-		await console.session.checkStatus(pythonSession, 'idle');
+		await session.checkStatus(pythonSession, 'active');
+		await session.checkStatus(pythonSession, 'idle');
 
 		// Restart Python session and confirm state returns to active --> idle
-		await console.session.restart(pythonSession, false);
-		await console.session.checkStatus(pythonSession, 'active');
-		await console.session.checkStatus(pythonSession, 'idle');
+		await session.restart(pythonSession, false);
+		await session.checkStatus(pythonSession, 'active');
+		await session.checkStatus(pythonSession, 'idle');
 
 		// Start R session
 		await app.workbench.console.startSession({ ...rSession, waitForReady: false });
 
 		// Verify R session transitions from active --> idle while Python session remains idle
-		await console.session.checkStatus(rSession, 'active');
-		await console.session.checkStatus(rSession, 'idle');
-		await console.session.checkStatus(pythonSession, 'idle');
+		await session.checkStatus(rSession, 'active');
+		await session.checkStatus(rSession, 'idle');
+		await session.checkStatus(pythonSession, 'idle');
 
 		// Shutdown Python session, verify Python transitions to disconnected while R remains idle
-		await console.session.shutdown(pythonSession, false);
-		await console.session.checkStatus(pythonSession, 'disconnected');
-		await console.session.checkStatus(rSession, 'idle');
+		await session.shutdown(pythonSession, false);
+		await session.checkStatus(pythonSession, 'disconnected');
+		await session.checkStatus(rSession, 'idle');
 
 		// Restart R session, verify R to returns to active --> idle and Python remains disconnected
-		await console.session.restart(rSession, false);
-		await console.session.checkStatus(rSession, 'active');
-		await console.session.checkStatus(rSession, 'idle');
-		await console.session.checkStatus(pythonSession, 'disconnected');
+		await session.restart(rSession, false);
+		await session.checkStatus(rSession, 'active');
+		await session.checkStatus(rSession, 'idle');
+		await session.checkStatus(pythonSession, 'disconnected');
 
 		// Shutdown R, verify both Python and R in disconnected state
-		await console.session.shutdown(rSession, false);
-		await console.session.checkStatus(rSession, 'disconnected');
-		await console.session.checkStatus(pythonSession, 'disconnected');
+		await session.shutdown(rSession, false);
+		await session.checkStatus(rSession, 'disconnected');
+		await session.checkStatus(pythonSession, 'disconnected');
 	});
 
 	test('Validate session state displays as active when executing code', async function ({ app }) {
+		const session = app.workbench.session;
 		const console = app.workbench.console;
 
 		// Ensure sessions exist and are idle
-		await console.session.ensureStartedAndIdle(pythonSession);
-		await console.session.ensureStartedAndIdle(rSession);
+		await session.ensureStartedAndIdle(pythonSession);
+		await session.ensureStartedAndIdle(rSession);
 
 		// Verify Python session transitions to active when executing code
-		await console.session.select(pythonSession);
+		await session.select(pythonSession);
 		await console.typeToConsole('import time', true);
 		await console.typeToConsole('time.sleep(1)', true);
-		await console.session.checkStatus(pythonSession, 'active');
-		await console.session.checkStatus(pythonSession, 'idle');
+		await session.checkStatus(pythonSession, 'active');
+		await session.checkStatus(pythonSession, 'idle');
 
 		// Verify R session transitions to active when executing code
-		await console.session.select(rSession);
+		await session.select(rSession);
 		await console.typeToConsole('Sys.sleep(1)', true);
-		await console.session.checkStatus(rSession, 'active');
-		await console.session.checkStatus(rSession, 'idle');
+		await session.checkStatus(rSession, 'active');
+		await session.checkStatus(rSession, 'idle');
 	});
 
 	test('Validate metadata between sessions', {
 		annotation: [
 			{ type: 'issue', description: 'https://github.com/posit-dev/positron/issues/6389' }]
 	}, async function ({ app }) {
-		const console = app.workbench.console;
+		const session = app.workbench.session;
 
 		// Ensure sessions exist and are idle
-		await console.session.ensureStartedAndIdle(pythonSession);
-		await console.session.ensureStartedAndIdle(rSession);
+		await session.ensureStartedAndIdle(pythonSession);
+		await session.ensureStartedAndIdle(rSession);
 
 		// Verify Python session metadata
-		await console.session.checkMetadata({ ...pythonSession, state: 'idle' });
-		await console.session.checkMetadata({ ...rSession, state: 'idle' });
+		await session.checkMetadata({ ...pythonSession, state: 'idle' });
+		await session.checkMetadata({ ...rSession, state: 'idle' });
 
 		// Shutdown Python session and verify metadata
-		await console.session.shutdown(pythonSession);
-		await console.session.checkMetadata({ ...pythonSession, state: 'exited' });
+		await session.shutdown(pythonSession);
+		await session.checkMetadata({ ...pythonSession, state: 'exited' });
 
 		// Shutdown R session and verify metadata
-		await console.session.shutdown(rSession);
-		await console.session.checkMetadata({ ...rSession, state: 'exited' });
+		await session.shutdown(rSession);
+		await session.checkMetadata({ ...rSession, state: 'exited' });
 	});
 
 	test('Validate variables between sessions', {
 		tag: [tags.VARIABLES]
 	}, async function ({ app }) {
+		const session = app.workbench.session;
 		const console = app.workbench.console;
 		const variables = app.workbench.variables;
 
 		// Ensure sessions exist and are idle
-		await console.session.ensureStartedAndIdle(pythonSession);
-		await console.session.ensureStartedAndIdle(rSession);
+		await session.ensureStartedAndIdle(pythonSession);
+		await session.ensureStartedAndIdle(rSession);
 
 		// Set and verify variables in Python
-		await console.session.select(pythonSession);
+		await session.select(pythonSession);
 		await console.typeToConsole('x = 1', true);
 		await console.typeToConsole('y = 2', true);
 		await variables.checkRuntime(pythonSession);
@@ -135,7 +137,7 @@ test.describe('Console: Sessions', {
 		await variables.checkVariableValue('y', '2');
 
 		// Set and verify variables in R
-		await console.session.select(rSession);
+		await session.select(rSession);
 		await console.typeToConsole('x <- 3', true);
 		await console.typeToConsole('z <- 4', true);
 		await variables.checkRuntime(rSession);
@@ -143,14 +145,14 @@ test.describe('Console: Sessions', {
 		await variables.checkVariableValue('z', '4');
 
 		// Switch back to Python, update variables, and verify
-		await console.session.select(pythonSession);
+		await session.select(pythonSession);
 		await console.typeToConsole('x = 0', true);
 		await variables.checkRuntime(pythonSession);
 		await variables.checkVariableValue('x', '0');
 		await variables.checkVariableValue('y', '2');
 
 		// Switch back to R, verify variables remain unchanged
-		await console.session.select(rSession);
+		await session.select(rSession);
 		await variables.checkRuntime(rSession);
 		await variables.checkVariableValue('x', '3');
 		await variables.checkVariableValue('z', '4');
@@ -161,39 +163,39 @@ test.describe('Console: Sessions', {
 			{ type: 'issue', description: 'sessions are not correctly sorted atm. see line 174.' }
 		]
 	}, async function ({ app }) {
-		const console = app.workbench.console;
+		const session = app.workbench.session;
 		const interpreter = app.workbench.interpreterNew;
 
 		// Ensure sessions exist and are idle
-		await console.session.ensureStartedAndIdle(pythonSession);
-		await console.session.ensureStartedAndIdle(rSession);
+		await session.ensureStartedAndIdle(pythonSession);
+		await session.ensureStartedAndIdle(rSession);
 
 		// Get active sessions and verify they match the session picker: order matters!
-		let activeSessionsFromConsole = await console.session.getActiveSessions();
+		let activeSessionsFromConsole = await session.getActiveSessions();
 		let activeSessionsFromPicker = await interpreter.getActiveSessions();
 		expect(activeSessionsFromConsole).toStrictEqual(activeSessionsFromPicker);
 
 		// Shutdown Python session and verify active sessions
-		await console.session.shutdown(pythonSession);
-		activeSessionsFromConsole = await console.session.getActiveSessions();
+		await session.shutdown(pythonSession);
+		activeSessionsFromConsole = await session.getActiveSessions();
 		activeSessionsFromPicker = await interpreter.getActiveSessions();
 		expect(activeSessionsFromConsole).toStrictEqual(activeSessionsFromPicker);
 
 		// Shutdown R session and verify active sessions
-		await console.session.shutdown(rSession);
-		activeSessionsFromConsole = await console.session.getActiveSessions();
+		await session.shutdown(rSession);
+		activeSessionsFromConsole = await session.getActiveSessions();
 		activeSessionsFromPicker = await interpreter.getActiveSessions();
 		expect(activeSessionsFromConsole).toStrictEqual(activeSessionsFromPicker);
 
 		// Start Python session (again) and verify active sessions
-		await console.session.start(pythonSession);
-		activeSessionsFromConsole = await console.session.getActiveSessions();
+		await session.start(pythonSession);
+		activeSessionsFromConsole = await session.getActiveSessions();
 		activeSessionsFromPicker = await interpreter.getActiveSessions();
 		expect(activeSessionsFromConsole).toStrictEqual(activeSessionsFromPicker);
 
 		// Restart Python session and verify active sessions
-		await console.session.restart(pythonSession);
-		activeSessionsFromConsole = await console.session.getActiveSessions();
+		await session.restart(pythonSession);
+		activeSessionsFromConsole = await session.getActiveSessions();
 		activeSessionsFromPicker = await interpreter.getActiveSessions();
 		expect(activeSessionsFromConsole).toStrictEqual(activeSessionsFromPicker);
 	});
