+++ test/e2e/tests/top-action-bar/session-button.test.ts
@@ -3,7 +3,7 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { SessionName } from '../../infra';
+import { SessionName } from '../../pages/session';
 import { test, tags } from '../_test.setup';
 
 test.use({
@@ -30,12 +30,12 @@ test.describe('Top Action Bar: Session Button', {
 	test('Python - Verify session starts and displays as running', async function ({ app }) {
 		await app.workbench.console.startSession({ ...pythonSession, triggerMode: 'top-action-bar' });
 		await app.workbench.interpreterNew.verifySessionIsSelected(pythonSession);
-		await app.workbench.console.session.checkStatus(pythonSession, 'idle');
+		await app.workbench.session.checkStatus(pythonSession, 'idle');
 	});
 
 	test('R - Verify session starts and displays as running', async function ({ app }) {
 		await app.workbench.console.startSession({ ...rSession, triggerMode: 'top-action-bar' });
 		await app.workbench.interpreterNew.verifySessionIsSelected(rSession);
-		await app.workbench.console.session.checkStatus(rSession, 'idle');
+		await app.workbench.session.checkStatus(rSession, 'idle');
 	});
 });
